// src/app/Interfaces/ebilling-rules.ts
export interface EbillingRules {
  id: number;
  from1: number;
  from2: number;
  from3: number;
  from4: number;
  to1: number;
  to2: number;
  to3: number;
  to4: number;
  rate1: string;
  rate2: string;
  rate3: string;
  rate4: string;
  demandCharge: string;
  commercialRate: string;
  commercialDc: string;
  meterRent: string;
  vat: string;
}